hw4 bug fix

+++ hw4/hw4/puzzle/Board.java
@@ -58,8 +58,8 @@ public class Board implements WorldState {
             for (int j = 0; j < size(); j++) {
                 if (intBoard[i][j] != goal[i][j]) {
                     if (intBoard[i][j] != 0) {
-                        int norRow = (int) Math.ceil(intBoard[i][j] / size) - 1;
-                        int norCol = (int) intBoard[i][j] - norRow  * size - 1;
+                        int norRow = (int) Math.ceil(intBoard[i][j] / (double) size) - 1;
+                        int norCol = (int) intBoard[i][j] - norRow * size - 1;
                         distance = distance + Math.abs(norRow - i) + Math.abs(norCol - j);
                     }
                 }
